pyproject.toml
[project]
name = "ai-job-scout"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi>=0.135.1",
    "greenlet>=3.3.2",
    "openai-agents[litellm,sqlalchemy]>=0.10.5",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "scalar-fastapi>=1.8.0",
    "sqlmodel>=0.0.37",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
exclude = ["alembic"]